begin rewriting datasets
